Refactored repositories for new ID import and new parse_list method.
